requirements.txt
fake2db==0.5.3
Faker==0.7.11
mysql-connector-python-rf==2.2.1
pytest==3.2.2
pytest-html==1.16.0
